Moved MechanicProvider, CategoryProvider and GameProvider to a folder named applicationProviders
