+++ src/components/applicationProviders/CategoryProvider.js
