+++ src/components/searchPage/SearchPage.js
@@ -1,7 +1,7 @@
 import React, {useContext} from "react";
 import { SearchFilters } from "./SearchFilters.js";
 import { SearchList } from "./SearchList.js";
-import { GameContext } from "../GameProvider.js";
+import { GameContext } from "../applicationProviders/GameProvider.js";
 import "./Search.css";
 
 // Responsible for displaying the Search Filters & Search List components
